perf: store schemas in an object instead of an array of key/value
Currently the schemas are passed around with two variables:
- `schemas` of type Array<{[key: string]: SchemaObject>`
- `schemasRefPaths` which lists the keys

There are a few issues:
- `schemas` is wrongly typed as `SchemaObject[]` in the code
- Two related arguments need to be passed to many function accepting schemas
- To check if an item is already in `schemas`, `schemasRefPaths.includes(key)` is used
which is O(schemas.length)
- It is not clear that `schemasRefPaths` is actually used to avoid
infinite recursion.

This commit uses only one variable, `schemas`. To avoid infinite
recursion, `schemasRefPaths` is used only where necessary as an
internal variable, and is renamed to `pendingSchemaRefs`. It is also
passed by copy when modified, so that it's restored to its previous value once the recursive call is ended.

Lookup is now fast: O(1) when looking up `schemas`, and when avoiding
infinite recursion `pendingPaths` is looked up as well but it is a short
array that does not contain any more elements than needed.

Typing is also fixed.