feat(config): merge passed schemas, links, ... in createDocument
Merge the passed config's components one level deeper, so as to be able to
specify individual schemas that will not be overwritten by the generated document.

Fix #1244

BREAKING CHANGE:

Schemas, links, examples, responses, ... passed to `SwaggerModule.createDocument`
via `config.components` are no longer overwritten if definitions of the same kind
are present in the generated document.

They will be overwritten only if a definition of the same kind AND with the
same name is present.